requirements.txt
absl-py==0.9.0
boto3==1.12.24
botocore==1.15.24
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
docutils==0.15.2
filelock==3.0.12
future==0.18.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
grpcio==1.27.2
hydra-core==0.11.3
idna==2.9
imageio==2.8.0
jmespath==0.9.5
joblib==0.14.1
loguru==0.4.1
Markdown==3.2.1
numpy==1.18.1
oauthlib==3.1.0
omegaconf==1.4.1
packaging==20.4
pandas==1.0.1
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==0.8.5
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
scikit-learn==0.22.2
scipy==1.4.1
sentencepiece==0.1.85
six==1.14.0
tensorboard==2.1.1
test-tube==0.7.5
tokenizers==0.8.1rc1
torch==1.6.0
tqdm==4.42.1
transformers==3.0.2
urllib3==1.25.8
Werkzeug==1.0.0
